requirements.txt
discord.py==1.6.0
jishaku==1.20.0.220